requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
djangorestframework==3.15.2
django-sslserver==0.22
psycopg2==2.9.10
markdown2==2.5.1